Moved birth date validator to common helper.
